+++ saplings/circuits/src/components/forms/propose_circuit/index.scss
@@ -137,7 +137,7 @@
       color: var(--color-grey);
       border-bottom: 1px solid var(--color-light-grey);
       padding: 1rem 0 0.5rem 2rem;
-      background-image: url('../../images/search_icon.svg');
+      background-image: url('../../../images/search_icon.svg');
       background-position: 0.3rem 1rem;
       background-repeat: no-repeat;
       background-size: 1rem;
